# apps/wanaku-barn-backend/src/main/webui/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "AuditEvent" : {
        "type" : "object",
        "properties" : {
          "schema_version" : {
            "type" : "string"
          },
          "event_id" : {
            "type" : "string"
          },
          "stream_id" : {
            "type" : "string"
          },
          "sequence" : {
            "type" : "integer",
            "format" : "int64"
          },
          "category" : {
            "type" : "string"
          },
          "decision" : {
            "type" : "string"
          },
          "correlation_id" : {
            "type" : "string"
          },
          "request_id" : {
            "type" : "string"
          },
          "actor" : {
            "type" : "string"
          },
          "protocol" : {
            "type" : "string"
          },
          "operation" : {
            "type" : "string"
          },
          "target_type" : {
            "type" : "string"
          },
          "target" : {
            "type" : "string"
          },
          "reason_code" : {
            "type" : "string"
          },
          "explanation" : {
            "type" : "string"
          },
          "policy_revision" : {
            "type" : "string"
          },
          "response_status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "duration_ms" : {
            "type" : "integer",
            "format" : "int64"
          },
          "redaction" : {
            "$ref" : "#/components/schemas/RedactionMetadata"
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "coverage_complete" : {
            "type" : "boolean"
          },
          "dropped_events" : {
            "type" : "integer",
            "format" : "int64"
          },
          "timestamp" : {
            "type" : "string"
          }
        }
      },
...
      "BarnArchive" : {
        "type" : "object",
        "properties" : {
          "formatVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "schemaVersion" : {
            "type" : "integer",
            "format" : "int64"
          },
          "exportedAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "dataStores" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DataStoreRecord"
            }
          },
          "catalogVersions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CatalogVersionRecord"
            }
          },
          "versionCounters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "auditEvents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AuditEvent"
            }
          },
          "auditSequence" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "CatalogVersionRecord" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "activatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "checksum" : {
            "type" : "string"
          },
          "origin" : {
            "type" : "string"
          },
          "actor" : {
            "type" : "string"
          },
          "failureReason" : {
            "type" : "string"
          },
          "restoredFrom" : {
            "type" : "integer",
            "format" : "int64"
          },
          "dataStoreName" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string"
          },
          "labels" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "rejected" : {
            "type" : "boolean"
          }
        }
      },
      "DataStore" : {
        "type" : "object",
        "properties" : {
          "labels" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string"
          }
        }
      },
      "DataStoreRecord" : {
        "type" : "object",
        "properties" : {
          "labels" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "createdBy" : {
            "type" : "string"
          },
          "updatedBy" : {
            "type" : "string"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
...
      "RedactionMetadata" : {
        "type" : "object",
        "properties" : {
          "redacted_fields" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "payload_captured" : {
            "type" : "boolean"
          },
          "payload_truncated" : {
            "type" : "boolean"
          }
        }
      },
...
      "WanakuError" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "WanakuResponseBarnArchive" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "$ref" : "#/components/schemas/BarnArchive"
          }
        }
      },
...
      "WanakuResponseDataStoreRecord" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "$ref" : "#/components/schemas/DataStoreRecord"
          }
        }
      },
...
      "WanakuResponseInteger" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "WanakuResponseListDataStoreRecord" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DataStoreRecord"
            }
          }
        }
      },
...
      "WanakuResponseMapStringInteger" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        }
      },
...
      "WanakuResponseVoid" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : { }
        }
      }
    }
  },
  "paths" : {
...
    "/api/v1/data-store" : {
      "put" : {
        "parameters" : [ {
          "name" : "expectedRevision",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DataStore"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseVoid"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update",
        "tags" : [ "Data Stores Resource" ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseVoid"
                }
              }
            }
          }
        },
        "summary" : "Remove By Name",
        "tags" : [ "Data Stores Resource" ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "labelFilter",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "name",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseListDataStoreRecord"
                }
              }
            }
          }
        },
        "summary" : "List Or Get By Name",
        "tags" : [ "Data Stores Resource" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DataStore"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseDataStoreRecord"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Add",
        "tags" : [ "Data Stores Resource" ]
      }
    },
    "/api/v1/data-store/labels" : {
      "delete" : {
        "parameters" : [ {
          "name" : "labelExpression",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseInteger"
                }
              }
            }
          }
        },
        "summary" : "Remove If",
        "tags" : [ "Data Stores Resource" ]
      }
    },
    "/api/v1/data-store/{id}" : {
      "delete" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "expectedRevision",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "If-Match",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseVoid"
                }
              }
            }
          }
        },
        "summary" : "Remove By Id",
        "tags" : [ "Data Stores Resource" ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseDataStoreRecord"
                }
              }
            }
          }
        },
        "summary" : "Get By Id",
        "tags" : [ "Data Stores Resource" ]
      }
    },
...
    "/api/v1/management/export" : {
      "get" : {
        "parameters" : [ {
          "name" : "includeAudit",
          "in" : "query",
          "schema" : {
            "type" : "boolean",
            "default" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseBarnArchive"
                }
              }
            }
          }
        },
        "summary" : "Export",
        "tags" : [ "Backup Resource" ]
      }
    },
    "/api/v1/management/import" : {
      "post" : {
        "parameters" : [ {
          "name" : "replace",
          "in" : "query",
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BarnArchive"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseMapStringInteger"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Import Archive",
        "tags" : [ "Backup Resource" ]
      }
    },
...
  },
...
  "info" : {
    "title" : "wanaku-barn-backend API",
    "version" : "0.3.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8180",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8180",
    "description" : "Auto generated value"
  } ]
}